requirements.txt
Flask==2.3.3
paddlepaddle==2.5.1
paddleocr==2.7.0.3
Pillow==10.0.1
PyMuPDF==1.23.5
opencv-python==4.8.1.78
numpy==1.24.3
python-dateutil==2.8.2 